+++ src/FlowSensor.h
@@ -1,20 +1,21 @@
 #pragma once
 #include <Arduino.h>
+
 #include "PinDefinitions.h"
 
 class FlowSensor {
-    public:
-        void init();
-        void tick();
-        void reset();
-        void flowtickIsr();
-        unsigned int get();
+   public:
+    void init();
+    void tick();
+    void reset();
+    void flowtickIsr();
+    float get();
 
-    private:
-        unsigned long debitStart;
-        bool firstTick;
-        unsigned long debitTick;
-        float debitConstant = 2.6; // Konstanta flow sensor
+   private:
+    unsigned long debitStart;
+    bool firstTick;
+    unsigned long debitTick;
+    float debitConstant = 2.6;  // Konstanta flow sensor
 };
 
 extern void flowtickIsr();
